refactor: models accommodated in individual files
