+++ AcademicProgrammingApplication/models/__init__.py
@@ -1,0 +1,7 @@
+from .Class import Class
+from .Contract import Contract
+from .Program import Program
+from .Semester import Semester
+from .Subject import Subject
+from .Teacher import Teacher
+from .Viatic import Viatic
